v2.0.0 enhancements.
