+++ docker-compose.yaml
@@ -5,29 +5,31 @@ services:
       volumes:
         - ./app:/usr/src/app/
 
-  k8s_labels:
-    container_name: k8s_labels
+  ev_logger:
+    container_name: ev_logger
     build: .
     volumes:
       - ./app:/usr/src/app/
     
     # Default. Emits labels to stdout every 300 seconds and no identation.
-    command: k8s_labels.py
-
-    # Optional. Emits labels to stdout every 5 seconds and indents by 2 spaces.
-    # command: k8s_labels.py --indent 2 --interval 5
-
-    # Optional. Emits labels to a file every 5 seconds and no indentation.
-    # command: k8s_labels.py --interval 5
+    command: ev_logger.py
 
     environment:
       - TZ=Australia/Sydney
-      - APP_KUBERNETES_IO_INSTANCE=my-instance
-      - APP_KUBERNETES_IO_MANAGED-BY=my-team
-      - APP_KUBERNETES_IO_PART-OF=my-collection
-      - APP_KUBERNETES_IO_INVENTORY_0_NAME=my-app
-      - APP_KUBERNETES_IO_INVENTORY_0_VERSION=v1
-      - APP_KUBERNETES_IO_INVENTORY_0_COMPONENT=my-component
-      - APP_KUBERNETES_IO_INVENTORY_1_NAME=my-app2
-      - APP_KUBERNETES_IO_INVENTORY_1_VERSION=v2
-      - APP_KUBERNETES_IO_INVENTORY_1_COMPONENT=my-component2
+      - EV_LOGGER_PREFIX=APP_CONTOSO_MS                       # Mandatory. Prefix for the environment variables.
+      # - EV_LOGGER_INDENT=2                                  # Optional. JSON indentation. Default is 0.
+      # - EV_LOGGER_INTERVAL=300                              # Optional. Interval (sec) to log. Default is 300.
+      # - EV_LOGGER_INTERVAL=0                                # Example of logging only once.
+      # - EV_LOGGER_INTERVAL=60                               # Example of logging every minute.
+
+      # The remainder of the environment variables are examples of the labels that will be emitted.
+      # They all start with the prefix APP_CONTOSO_MS, which is defined by EV_LOGGER_PREFIX.
+      - APP_CONTOSO_MS_INSTANCE=my-instance
+      - APP_CONTOSO_MS_MANAGED-BY=my-team
+      - APP_CONTOSO_MS_PART-OF=my-collection
+      - APP_CONTOSO_MS_INVENTORY_0_NAME=my-app
+      - APP_CONTOSO_MS_INVENTORY_0_VERSION=v1
+      - APP_CONTOSO_MS_INVENTORY_0_COMPONENT=my-component
+      - APP_CONTOSO_MS_INVENTORY_1_NAME=my-app2
+      - APP_CONTOSO_MS_INVENTORY_1_VERSION=v2
+      - APP_CONTOSO_MS_INVENTORY_1_COMPONENT=my-component2
